chore(team): split mailbox mention helpers
