+++ src/team/manager.rs
@@ -6,6 +6,7 @@ mod context_artifacts;
 mod conversation;
 mod mailbox;
 mod mailbox_errors;
+mod mailbox_mentions;
 mod mailbox_payloads;
 mod mailbox_reply_obligations;
 mod mailbox_service;
